cleanup semantics for geometry, events, hazards

+++ packages/simulation/src/math/circle.rs
@@ -3,11 +3,11 @@ use crate::math::Point2;
 #[derive(Debug, Clone, PartialEq)]
 pub struct Circle {
     center: Point2,
-    radius: f32,
+    radius: f64,
 }
 
 impl Circle {
-    pub fn new(center: Point2, radius: f32) -> Self {
+    pub fn new(center: Point2, radius: f64) -> Self {
         Self {
             center,
             radius,
